cdk-backend/requirements.txt
aws-cdk-lib>=2.210.0
constructs>=10.0.0,<11.0.0
cdk-nag>=2.36.57
boto3>=1.40.5
aws-lambda-powertools[all]>=3.18.0
